Add source vocabulary client api
ignore '.tmuxinator.yml' file.

The '--parallel' flag increases the performance but does not show how
many tests passed, skipped, or failed. So remove the '--parallel' flag.

The SourceVocabulary API is stable. So no need for the
'SourceVocabulary' compilation flag.

The 'swift-url' dependency is added to create specifically the web URL
where the standard URL object can be used for files, websites, etc.

Library UMLSClientModel abstracts out the request and response model
objects. Therefore, add the library as a dependency to access the model
objects.